+++ packages/editor/src/store/test/actions.js
@@ -5,6 +5,8 @@ import {
 	replaceBlocks,
 	startTyping,
 	stopTyping,
+	enterFormattedText,
+	exitFormattedText,
 	fetchReusableBlocks,
 	saveReusableBlock,
 	deleteReusableBlock,
@@ -35,12 +37,6 @@ import {
 	removeBlocks,
 	removeBlock,
 	toggleBlockMode,
-	createNotice,
-	createSuccessNotice,
-	createInfoNotice,
-	createErrorNotice,
-	createWarningNotice,
-	removeNotice,
 	updateBlockListSettings,
 } from '../actions';
 
@@ -48,12 +44,10 @@ describe( 'actions', () => {
 	describe( 'setupEditor', () => {
 		it( 'should return the SETUP_EDITOR action', () => {
 			const post = {};
-			const autosave = {};
-			const result = setupEditor( post, autosave );
+			const result = setupEditor( post );
 			expect( result ).toEqual( {
 				type: 'SETUP_EDITOR',
 				post,
-				autosave,
 			} );
 		} );
 	} );
@@ -352,110 +346,18 @@ describe( 'actions', () => {
 		} );
 	} );
 
-	describe( 'createNotice', () => {
-		const status = 'status';
-		const content = <p>element</p>;
-		it( 'should return CREATE_NOTICE action when options is empty', () => {
-			const result = createNotice( status, content );
-			expect( result ).toMatchObject( {
-				type: 'CREATE_NOTICE',
-				notice: {
-					status,
-					content,
-					isDismissible: true,
-					id: expect.any( String ),
-				},
-			} );
-		} );
-		it( 'should return CREATE_NOTICE action when options is desined', () => {
-			const id = 'my-id';
-			const options = {
-				id,
-				isDismissible: false,
-			};
-			const result = createNotice( status, content, options );
-			expect( result ).toEqual( {
-				type: 'CREATE_NOTICE',
-				notice: {
-					id,
-					status,
-					content,
-					isDismissible: false,
-				},
+	describe( 'enterFormattedText', () => {
+		it( 'should return the ENTER_FORMATTED_TEXT action', () => {
+			expect( enterFormattedText() ).toEqual( {
+				type: 'ENTER_FORMATTED_TEXT',
 			} );
 		} );
 	} );
 
-	describe( 'createSuccessNotice', () => {
-		it( 'should return CREATE_NOTICE action', () => {
-			const content = <p>element</p>;
-			const result = createSuccessNotice( content );
-			expect( result ).toMatchObject( {
-				type: 'CREATE_NOTICE',
-				notice: {
-					status: 'success',
-					content,
-					isDismissible: true,
-					id: expect.any( String ),
-				},
-			} );
-		} );
-	} );
-
-	describe( 'createInfoNotice', () => {
-		it( 'should return CREATE_NOTICE action', () => {
-			const content = <p>element</p>;
-			const result = createInfoNotice( content );
-			expect( result ).toMatchObject( {
-				type: 'CREATE_NOTICE',
-				notice: {
-					status: 'info',
-					content,
-					isDismissible: true,
-					id: expect.any( String ),
-				},
-			} );
-		} );
-	} );
-
-	describe( 'createErrorNotice', () => {
-		it( 'should return CREATE_NOTICE action', () => {
-			const content = <p>element</p>;
-			const result = createErrorNotice( content );
-			expect( result ).toMatchObject( {
-				type: 'CREATE_NOTICE',
-				notice: {
-					status: 'error',
-					content,
-					isDismissible: true,
-					id: expect.any( String ),
-				},
-			} );
-		} );
-	} );
-
-	describe( 'createWarningNotice', () => {
-		it( 'should return CREATE_NOTICE action', () => {
-			const content = <p>element</p>;
-			const result = createWarningNotice( content );
-			expect( result ).toMatchObject( {
-				type: 'CREATE_NOTICE',
-				notice: {
-					status: 'warning',
-					content,
-					isDismissible: true,
-					id: expect.any( String ),
-				},
-			} );
-		} );
-	} );
-
-	describe( 'removeNotice', () => {
-		it( 'should return REMOVE_NOTICE actions', () => {
-			const noticeId = 'id';
-			expect( removeNotice( noticeId ) ).toEqual( {
-				type: 'REMOVE_NOTICE',
-				noticeId,
+	describe( 'exitFormattedText', () => {
+		it( 'should return the EXIT_FORMATTED_TEXT action', () => {
+			expect( exitFormattedText() ).toEqual( {
+				type: 'EXIT_FORMATTED_TEXT',
 			} );
 		} );
 	} );