+++ packages/edit-post/CHANGELOG.md
@@ -1,6 +1,25 @@
-## 1.0.4 (Unreleased)
+## 2.0.0 (2018-10-29)
 
-## Bug Fixes
+### Breaking Changes
+
+- `setActiveMetaBoxLocations` action (`core/edit-post`) has been removed.
+- `initializeMetaBoxState` action (`core/edit-post`) has been removed.
+- `initializeEditor` no longer returns an object. Use the `setActiveMetaBoxLocations` action (`core/edit-post`) in place of the existing object's `initializeMetaBoxes` function.
+- `setMetaBoxSavedData` action (`core/edit-post`) has been removed.
+- `getMetaBoxes` selector (`core/edit-post`) has been removed. Use `getActiveMetaBoxLocations` selector (`core/edit-post`) instead.
+- `getMetaBox` selector (`core/edit-post`) has been removed. Use `isMetaBoxLocationActive` selector (`core/edit-post`) instead.
+
+### Polish
+
+- Add the editor styles support's wrapper className.
+
+### Bug Fixes
+
+- Hide pinned plugins and block traversal tool from header on mobile.
+
+## 1.0.4 (2018-10-22)
+
+### Bug Fixes
 
 - Fix fullscreen mode toggle.